Crate libb2b

A simple, but relatively extensible peer-to-peer chat system. Makes heavy use of tokio for async programming. Has the base functionality to enable shortest path algorithms (e.g., Dijkstra’s).

The basic is that a given peer is composed of two parts:

A Client

A Server

Re-exports

pub use tracker::Tracker;

pub use cmd::Bing2BingCommand;

Modules

client

cmd

connection

parse

peer

peer_map

server

tracker

util

Structs

Client
A Client is the way that a user (i.e., a user of our crate) interacts with a Server, and thus the rest of the network..

Connection
A Connection handles reading/writing to the network.

Server
The “server” side of the P2P chat application. A server is primarily focused around network related activity and manages most everything related to the protocol itself. This includes receiving commands over the network, processing them, and sending commands out to the network. The server also receives messages from its corresponding Client which is what the end user will be interacting with.

Enums

Bing2BingFrame

ClientServerMessage

PeerControlMessage

Functions

init
Initializes a new “peer.” I.e., creates a Client and a Server pair. Upon return, the Server will be ready to be Server::start()ed.

Type Definitions

Bing2BingError

BingLengthDelimitedCodec

ClientRxChannel

ClientTxChannel

Framed

PeerRxChannel

PeerTxChannel

ServerRxChannel

ServerTxChannel