Module libb2b::cmd

Modules

announce

broadcast

deliver

extension

ping

register

say

whisper

Structs

Announce
The Announce command is propagated through the network to provide peers knowledge about the network topography. I.e., this is how peers let each other know who they are connected to.

Broadcast
The Broadcast command delivers data (a Bing2BingFrame::Bulk) to all connected peers.

Deliver
Deliver data Bing2BingFrame::Bulk to a specific destination (peer).

Extension
This command serves as a mechanism to enable extensions to the protocol. It is esssentially a wrapper around:

Ping
A simple command that let’s peers test latency between each other.

Register
This command is sent to the tracker to help a peer boostrap itself.

Say
This command allows for propagation of normal chat messages through out the network.

Whisper
This command allows for direct messaging between two peers. The idea is that peers should forward this message via the shortest path to the target.

Enums

Bing2BingCommand