Enum libb2b::cmd::Bing2BingCommand

pub enum Bing2BingCommand {
    Broadcast(Broadcast),
    Ping(Ping),
    Register(Register),
    Say(Say),
    Deliver(Deliver),
    Announce(Announce),
    Whisper(Whisper),
    Extension(Extension),
    Unknown,
}

Variants

Broadcast(Broadcast)

Ping(Ping)

Register(Register)

Say(Say)

Deliver(Deliver)

Announce(Announce)

Whisper(Whisper)

Extension(Extension)

Unknown

Implementations

impl Bing2BingCommand

pub(crate) fn from_frame(
frame: Bing2BingFrame
) -> Result<Bing2BingCommand, Bing2BingError>

pub fn into_frame(cmd: Bing2BingCommand) -> Bing2BingFrame

POINTS AVAILABLE There is a way to refactor things such that we could do a call like Bing2BingCommand::into_frame(cmd) instead of having to call cmd.into_frame() directly. This would give us some benefits with respect to ergonomics (we wouldn’t have to have the underlying cmd struct) fully typed.

pub(crate) fn check_duplicate(&self, processed_commands: &TtlMap<bool>) -> bool

Checks to make sure that this Bing2BingCommand hasn’t already been processed. This helps us ensure that we don’t start an infinite loop.

pub(crate) fn set_processed(&self, processed_commands: &TtlMap<bool>)