Function libb2b::server::start_announce

fn start_announce(
    name: String, 
    ip_address: String, 
    port: u64, 
    peer_map: &PeerMap, 
    next_sequence_number: SequenceNumberGenerator, 
    num_incoming_conns: ConnectionCounter, 
    max_incoming_conns: u64
)

POINTS AVAILABLE Right now, this method will announce the peer to the rest of the network every 5 seconds. As part of this announcement, the peer will transmit the name of the city it’s in, as well as lat and longitude. It would be nice to have this be configurable instead of hard coded as it currently is.