Struct libb2b::Server

pub struct Server {
    listener: TcpListener,
    sequence_numbers: SequenceNumberGenerator,
    name: String,
    ip_address: String,
    port: u64,
    num_incoming_conns: ConnectionCounter,
    client_tx: ClientTxChannel,
    rx: ServerRxChannel,
}

The “server” side of the P2P chat application. A server is primarily focused around network related activity and manages most everything related to the protocol itself. This includes receiving commands over the network, processing them, and sending commands out to the network. The server also receives messages from its corresponding Client which is what the end user will be interacting with.

Fields
listener: TcpListenersequence_numbers: SequenceNumberGeneratorname: Stringip_address: Stringport: u64num_incoming_conns: ConnectionCounterclient_tx: ClientTxChannelrx: ServerRxChannel
Implementations

impl Server

pub async fn new(
name: &str,
bind_address: &str,
port: &str,
client_tx: ClientTxChannel,
rx: ServerRxChannel
) -> Result<Self, Bing2BingError>

pub async fn listen(
&self,
peer_map: &PeerMap,
client_tx: ClientTxChannel
) -> Result<(), Bing2BingError>

Begin listening for inbound connections.

pub async fn handle_connection(
peers: &PeerMap,
adjacency_list: TtlMap<PeerData>,
processed_commands: TtlMap<bool>,
stream: TcpStream,
addr: SocketAddr,
client_tx: ClientTxChannel
) -> Result<(), Bing2BingError>

Handles an incomming connection. I.e., another peer that has initiated a connection with us. In particular, this method reads command frames from a Connection, checks to make sure that the command hasn’t already been processed, and if it hasn’t, processes the command. This method will also pass relevant ClientServerMessages up to a Client for further use.

pub async fn say(
peer_map: &PeerMap,
from: String,
message: String,
sequence_number: u64
)

Convienence function that broadcasts a say message. This is useful for handling messages that are coming in from the associated Client. I.e., our user wants to say something.

fn next_sequence_number(&self) -> u64

Convienence function that gets the next sequence number for a message originating from this peer.

pub async fn start(
&self,
tracker_ip: &str,
tracker_port: &str,
max_incoming_connections: u64
) -> Result<(), Bing2BingError>

Starts the server. This is primarily three steps:

We want to register with the tracker.

We want to connect to peers that we get back from the tracker.

We want to start listening for incoming connections from other peers.

We want to start announcing our neighborhood to others.

fn client_message_handler(&self, peer_map: &PeerMap, rx: ServerRxChannel)

This method handles messages that come in from the associated Client

fn parse_register_response(
&self,
response: Bing2BingFrame
) -> Result<Vec<(String, String, String)>, Bing2BingError>

pub(crate) fn connect_to_peer(
peer_map: &PeerMap,
peer_name: String,
ip_address: String,
port: String
)